restructured input panel to comply with the MVC pattern
