+++ health/src/main/java/com/health/gui/GUImain.java
@@ -13,6 +13,7 @@ import javax.swing.UIManager.LookAndFeelInfo;
 import javax.swing.UnsupportedLookAndFeelException;
 import javax.swing.plaf.basic.BasicTabbedPaneUI;
 
+import com.health.control.InputController;
 import com.health.gui.input.VInputPanel;
 import com.health.gui.output.VOutputPanel;
 import com.health.gui.script.VScriptPanel;
@@ -104,9 +105,11 @@ public class GUImain extends JFrame implements UserInterface {
 
     private void createTabbedPane() {
         tabbedPane = new JTabbedPane();
-        VidneyPanel inputPanel = new VInputPanel();
+        VInputPanel inputPanel = new VInputPanel();
         VidneyPanel scriptPanel = new VScriptPanel();
         VidneyPanel outputPanel = new VOutputPanel();
+        
+        InputController inputController = new InputController(inputPanel);
 
         addTab("Step 1: Input", inputPanel);
         addTab("Step 2: Script", scriptPanel);
